setup prettier and fix fprmats
