+++ src/containers/Auth/index.js
@@ -1,0 +1,73 @@
+import { Form, Input, Button, Checkbox } from "antd";
+import { UserOutlined, LockOutlined } from "@ant-design/icons";
+import { useSelector, useDispatch } from "react-redux";
+import { useHistory } from "react-router-dom";
+
+import { signin } from "../../actions/auth";
+import "./auth.css";
+
+export default function Auth() {
+    const dispatch = useDispatch();
+    const history = useHistory();
+    const authMessage = useSelector((state) => state.auth);
+
+    const onFinish = (values) => {
+        dispatch(signin(values, history));
+        console.log(authMessage);
+    };
+    return (
+        <div className="login-container">
+            <Form
+                name="normal_login"
+                className="login-form"
+                initialValues={{
+                    remember: true,
+                }}
+                onFinish={onFinish}
+            >
+                <Form.Item
+                    name="username"
+                    rules={[
+                        {
+                            required: true,
+                            message: "Please input your Username!",
+                        },
+                    ]}
+                >
+                    <Input
+                        prefix={
+                            <UserOutlined className="site-form-item-icon" />
+                        }
+                        placeholder="Username"
+                    />
+                </Form.Item>
+                <Form.Item
+                    name="password"
+                    rules={[
+                        {
+                            required: true,
+                            message: "Please input your Password!",
+                        },
+                    ]}
+                >
+                    <Input
+                        prefix={
+                            <LockOutlined className="site-form-item-icon" />
+                        }
+                        type="password"
+                        placeholder="Password"
+                    />
+                </Form.Item>
+                <Form.Item>
+                    <Button
+                        type="primary"
+                        htmlType="submit"
+                        className="login-form-button"
+                    >
+                        Log in
+                    </Button>
+                </Form.Item>
+            </Form>
+        </div>
+    );
+}
